+++ haskell/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 0.1.0.0 -- YYYY-mm-dd
 
-* First version. Released on an unsuspecting world.
+* Added lexer for initial tokens from the book